plugins/render/*: Updated all x86 asm renderers to use the new MDP_Render_t.destPitch and MDP_Render_t.srcPitch parameters.

+++ src/gens/plugins/render/scanline/mdp_render_scanline_x86.h
@@ -37,11 +37,11 @@ void mdp_render_scanline_16_x86_mmx(uint16_t *destScreen, uint16_t *mdScreen,
 				    int destPitch, int srcPitch,
 				    int width, int height);
 
-void mdp_render_scanline_32_x86(uint16_t *destScreen, uint16_t *mdScreen,
+void mdp_render_scanline_32_x86(uint32_t *destScreen, uint32_t *mdScreen,
 				int destPitch, int srcPitch,
 				int width, int height);
 
-void mdp_render_scanline_32_x86_mmx(uint16_t *destScreen, uint16_t *mdScreen,
+void mdp_render_scanline_32_x86_mmx(uint32_t *destScreen, uint32_t *mdScreen,
 				    int destPitch, int srcPitch,
 				    int width, int height);
